material alert dialog items options, and intents

+++ app/src/main/java/com/godiapper/stores/ui/view/MainActivity.kt
@@ -1,7 +1,11 @@
 package com.godiapper.stores.ui.view
 
+import android.content.DialogInterface
+import android.content.Intent
+import android.net.Uri
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.widget.Toast
 import androidx.recyclerview.widget.GridLayoutManager
 import com.godiapper.stores.R
 import com.godiapper.stores.adapter.StoreAdapter
@@ -9,6 +13,7 @@ import com.godiapper.stores.core.OnClickListener
 import com.godiapper.stores.core.StoreApplication
 import com.godiapper.stores.core.StoreEntity
 import com.godiapper.stores.databinding.ActivityMainBinding
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import org.jetbrains.anko.doAsync
 import org.jetbrains.anko.uiThread
 
@@ -56,7 +61,7 @@ class MainActivity : AppCompatActivity(), OnClickListener, MainAux {
 
     private fun setupReciclerView() {
         mAdapter = StoreAdapter(mutableListOf(),this)
-        mGridLayout = GridLayoutManager(this,2)
+        mGridLayout = GridLayoutManager(this,resources.getInteger(R.integer.main_columns))
         getStores()
 
         mbinding.recyclerView.apply {
@@ -88,18 +93,71 @@ class MainActivity : AppCompatActivity(), OnClickListener, MainAux {
         doAsync {
             StoreApplication.database.storeDao().updateStore(storeEntity)
             uiThread {
-                mAdapter.update(storeEntity)
+                updateStore(storeEntity)
             }
         }
     }
 
     override fun onDeleteStore(storeEntity: StoreEntity) {
-        doAsync {
-            StoreApplication.database.storeDao().deleteStore(storeEntity)
-            uiThread {
-                mAdapter.delete(storeEntity)
-            }
+        val items = resources.getStringArray(R.array.array_option_item)
+
+        MaterialAlertDialogBuilder(this)
+            .setTitle(R.string.dialog_options_title)
+            .setItems(items,{dialogInterface, i ->
+                when(i){
+                    0 -> confirmDelete(storeEntity)
+
+                    1 -> dial(storeEntity.phone)
+
+                    2 -> goToWebsite(storeEntity.website)
+                }
+            })
+            .show()
+    }
+
+    private fun confirmDelete(storeEntity: StoreEntity){
+        MaterialAlertDialogBuilder(this)
+            .setTitle(R.string.dialog_delete_title)
+            .setPositiveButton(R.string.dialog_delete_confirmar,{ dialogInterface, i ->
+                doAsync {
+                    StoreApplication.database.storeDao().deleteStore(storeEntity)
+                    uiThread {
+                        mAdapter.delete(storeEntity)
+                    }
+                }
+            })
+            .setNegativeButton(R.string.dialog_delete_cancel,null)
+            .show()
+    }
+
+    private fun dial(phone: String) {
+        val callIntent = Intent().apply {
+            action = Intent.ACTION_DIAL
+            data = Uri.parse("tel:$phone")
         }
+
+        startIntent(callIntent)
+    }
+
+    private fun goToWebsite(website: String){
+        if(website.isEmpty()){
+            Toast.makeText(this,R.string.main_error_no_website, Toast.LENGTH_LONG).show()
+        } else{
+            val websiteIntent = Intent().apply {
+                action = Intent.ACTION_VIEW
+                data = Uri.parse(website)
+            }
+
+            startIntent(websiteIntent)
+        }
+
+    }
+
+    private fun startIntent(intent: Intent){
+        if (intent.resolveActivity(packageManager) != null)
+            startActivity(intent)
+        else
+            Toast.makeText(this,R.string.main_error_no_resolve, Toast.LENGTH_LONG).show()
     }
 
     override fun hideFab(isVisible: Boolean) {
